# server/swagger.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
  description: API documentation for the marketplace platform.
servers:
  - url: http://157.245.187.246:5000
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /api/users/{userId}:
    get:
      tags: [Users]
      summary: Get a public profile by userId
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: The user's public ID
      responses:
        '200':
          description: Public profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error fetching user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/products:
    post:
      tags: [Products]
      summary: Add a new product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, price, originAddress, type, quantity]
              properties:
                title:
                  type: string
                price:
                  type: number
                  minimum: 0
                originAddress:
                  type: string
                type:
                  type: string
                quantity:
                  type: integer
                  minimum: 1
                description:
                  type: string
                comment:
                  type: string
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 5
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Restricted user or unverified user cannot add products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error adding product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Products]
      summary: Get all products (with pagination & filters)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
      responses:
        '200':
          description: Products list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '500':
          description: Server error fetching products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/products/{productId}:
    get:
      tags: [Products]
      summary: Get a single product
      parameters:
        - in: path
          name: productId
          schema:
            type: string
          required: true
          description: The product's public ID or MongoDB ObjectId
      responses:
        '200':
          description: Product retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error fetching product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: Get all users (admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (not an admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error fetching users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/users/{userId}:
    get:
      tags: [Admin]
      summary: Get a single user profile (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: The user's public ID
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (not an admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error fetching user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/users/{userId}/restrict:
    post:
      tags: [Admin]
      summary: Restrict or unrestrict a user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: The user's public ID
      responses:
        '200':
          description: User restricted or unrestricted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (not an admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error restricting user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/transactions:
    get:
      tags: [Admin]
      summary: Get all transactions (admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (not an admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error fetching transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/users/suspicious:
    get:
      tags: [Admin]
      summary: Get suspicious users (admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of suspicious users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  suspiciousUsers:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (not an admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error fetching suspicious users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/verifications/pending:
    get:
      tags: [Admin]
      summary: Get pending government ID verifications (admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users with pending ID verifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  pending:
                    type: array
                    items:
                      type: object
                      properties:
                        userId:
                          type: string
                        fullName:
                          type: string
                        email:
                          type: string
                        govIdFront:
                          type: string
                        govIdBack:
                          type: string
                        govIdStatus:
                          type: string
                          enum: ['not_uploaded', 'pending', 'approved', 'rejected']
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (not an admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error fetching pending verifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/verify/{userId}:
    patch:
      tags: [Admin]
      summary: Approve or reject a user's government ID verification (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: The user's public ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: ['approve', 'reject']
                  description: Action to take on the verification
              required: ['action']
      responses:
        '200':
          description: Verification processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  user:
                    type: object
                    properties:
                      userId:
                        type: string
                      fullName:
                        type: string
                      email:
                        type: string
                      verified:
                        type: boolean
                      govIdStatus:
                        type: string
                        enum: ['not_uploaded', 'pending', 'approved', 'rejected']
        '400':
          description: Invalid action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (not an admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error processing ID verification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/messages:
    get:
      tags: [Admin]
      summary: Get all contact form messages (admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (not an admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error fetching messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/messages/{messageId}/reply:
    post:
      tags: [Admin]
      summary: Reply to a contact form message (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: messageId
          required: true
          schema:
            type: string
          description: The message's MongoDB ObjectId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reply:
                  type: string
                  description: The admin's reply to the message
              required: ['reply']
      responses:
        '200':
          description: Reply sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      messageId:
                        type: string
                      name:
                        type: string
                      email:
                        type: string
                      subject:
                        type: string
                      reply:
                        type: string
                      status:
                        type: string
                        enum: ['pending', 'replied']
                      createdAt:
                        type: string
                        format: date-time
                      repliedAt:
                        type: string
                        format: date-time
                        nullable: true
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (not an admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error replying to message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/make-admin/{userId}:
    post:
      tags: [Admin]
      summary: Promote a user to admin (dev only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User promoted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error promoting admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/platform-fees:
    get:
      tags: [Admin]
      summary: Get total platform fees
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Total platform fees retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  totalFees:
                    type: number
                  feeCount:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error fetching platform fees
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        userId:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        username:
          type: string
        location:
          type: string
        profilePic:
          type: string
        govIdFront:
          type: string
        govIdBack:
          type: string
        verified:
          type: boolean
        registrationDate:
          type: string
          format: date-time
        rank:
          type: number
        customerRating:
          type: number
        balance:
          type: number
        pendingBalance:
          type: number
        isAdmin:
          type: boolean
        isRestricted:
          type: boolean
        govIdStatus:
          type: string
          enum: [not_uploaded, pending, approved, rejected]
        postedProducts:
          type: array
          items:
            type: string
        savedProducts:
          type: array
          items:
            type: string
        boughtProducts:
          type: array
          items:
            type: string
        soldProducts:
          type: array
          items:
            type: string
        closeCustomers:
          type: array
          items:
            type: string
        transactionHistory:
          type: array
          items:
            type: string
...
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: '#/components/schemas/User'
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: string
...
    ProductResponse:
      type: object
      properties:
        success:
          type: boolean
        product:
          $ref: '#/components/schemas/Product'
    ProductListResponse:
      type: object
      properties:
        success:
          type: boolean
        page:
          type: integer
        pages:
          type: integer
        total:
          type: integer
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
...